Add benchmarks for different Remap error handling methods
